+++ tasks.py
@@ -134,7 +134,7 @@ def test(c, target=None):
         c.run(f"{PYTHON} -m pytest tests/windows")
 
     if target:
-        modules = ['local', 'model', 'plugin.py', 'consts.py', 'humbledownloader.py', 'webservice.py', 'settings.py', 'library.py']
+        modules = ['local', 'model', 'plugin.py', 'consts.py', 'humbledownloader.py', 'webservice.py', 'settings.py', 'library.py', 'active_month_resolver.py']
         os.environ['MYPYPATH'] = str(Path(target) / THIRD_PARTY_RELATIVE_DEST)
         modules_full_path = [str(Path(target) / mod) for mod in modules]
         print(f'running mypy check for {str(Path(target))} directory')